Modification Class Expression

+++ Partie 2/Expression.java
@@ -1,4 +1,4 @@
 abstract class Expression {
     // Méthode abstraite pour retourner la valeur de l'expression
-    public abstract double valeur();
+    public abstract int valeur();
 }
